+++ src/backend/joanie/core/enums.py
@@ -59,6 +59,9 @@ CATALOG_VISIBILITY_CHOICES = (
 
 ORDER_STATE_DRAFT = "draft"  # order has been created
 ORDER_STATE_ASSIGNED = "assigned"  # order has been assigned to an organization
+ORDER_STATE_TO_OWN = (
+    "to_own"  # order is paid with batch order and doesn't have owner yet
+)
 ORDER_STATE_TO_SAVE_PAYMENT_METHOD = (
     "to_save_payment_method"  # order needs a payment method
 )
@@ -105,6 +108,12 @@ ORDER_STATE_CHOICES = (
         ORDER_STATE_REFUNDED,
         pgettext_lazy("As in: the order payments are refunded", "Refunded"),
     ),
+    (
+        ORDER_STATE_TO_OWN,
+        pgettext_lazy(
+            "As in: the order is paid through batch order but without owner", "To own"
+        ),
+    ),
 )
 ORDER_STATE_ALLOW_ENROLLMENT = (
     ORDER_STATE_COMPLETED,
@@ -122,6 +131,7 @@ ORDER_INACTIVE_STATES = (
     ORDER_STATE_REFUNDING,
     ORDER_STATE_REFUNDED,
 )
+ORDER_STATES_VOUCHER_CLAIMABLE = (*ORDER_INACTIVE_STATES, ORDER_STATE_TO_OWN)
 
 ENROLLMENT_STATE_SET = "set"
 ENROLLMENT_STATE_FAILED = "failed"
@@ -234,3 +244,22 @@ PAYMENT_STATE_CHOICES = (
 # Course offers
 COURSE_OFFER_PAID = "paid"
 COURSE_OFFER_FREE = "free"
+
+
+BATCH_ORDER_STATE_DRAFT = "draft"
+BATCH_ORDER_STATE_ASSIGNED = "assigned"
+BATCH_ORDER_STATE_TO_SIGN = "to_sign"
+BATCH_ORDER_STATE_SIGNING = "signing"
+BATCH_ORDER_STATE_PENDING = "pending"
+BATCH_ORDER_STATE_FAILED_PAYMENT = "failed_payment"
+BATCH_ORDER_STATE_COMPLETED = "completed"
+
+BATCH_ORDER_STATE_CHOICES = (
+    (BATCH_ORDER_STATE_DRAFT, _("Draft")),
+    (BATCH_ORDER_STATE_ASSIGNED, _("Assigned")),
+    (BATCH_ORDER_STATE_TO_SIGN, _("To sign")),
+    (BATCH_ORDER_STATE_SIGNING, _("Signing")),
+    (BATCH_ORDER_STATE_PENDING, _("Pending")),
+    (BATCH_ORDER_STATE_FAILED_PAYMENT, _("Failed payment")),
+    (BATCH_ORDER_STATE_COMPLETED, _("Completed")),
+)